feat(registry): cover legacy component installs
