+++ app/(docs)/docs/discloseimage/page.tsx
@@ -1,9 +1,10 @@
-import DiscloseImage from "@/components/spectrumui/discloseimage";
 import React from "react";
 import { PageTemplate } from "@/app/(docs)/docs/components/page-template";
+import PreviewCodeCard from "@/app/(docs)/docs/components/preview-code-card";
 import { Metadata } from "next";
 import { baseMetadata } from "@/app/(docs)/layout-parts/base-metadata";
 import { SEOWrapper } from "@/app/(docs)/docs/components/seo-wrapper";
+import DiscloseImageDemo from "./disclose-image-demo";
 
 export const metadata: Metadata = baseMetadata({
   title: "Disclose Image",
@@ -43,14 +44,13 @@ const page = () => {
         title="Disclose Image"
         description="An image that reveals its content behind animated sliding door panels."
       >
-        <div className="mt-6">
-          <DiscloseImage
-            alt="A beautiful image"
-            vertical
-            doorClassName="bg-green-500"
-            src="https://plus.unsplash.com/premium_vector-1689096860582-07eee139f9f1?bg=FFFFFF&w=800&auto=format&fit=crop&q=100&ixlib=rb-4.0.3"
-          />
-        </div>
+        <PreviewCodeCard
+          path="app/(docs)/docs/discloseimage/disclose-image-demo.tsx"
+          cli="@spectrumui/disclose-image"
+          installCodePath="components/spectrumui/discloseimage.tsx"
+        >
+          <DiscloseImageDemo />
+        </PreviewCodeCard>
       </PageTemplate>
     </SEOWrapper>
   );
